[Feat/#48]: models 생성
