+++ src/components/modal/CommentItem.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { postAPI } from '@/api/postAPI';
-import { Comment as IComment } from '@/types/api/response';
+import { postAPI } from '@/api/requests/postAPI';
+import { Comment as IComment } from '@/models/post.model';
 import { formatDate } from '@/utils/format';
 import { LEVEL } from '@/constants/level';
 import { ICONS } from '../../constants/assets';
