+++ src/models/common.model.ts
@@ -1,0 +1,4 @@
+export interface CommonRes {
+  isSuccess: boolean;
+  message: string;
+}
